# [Generative Score API] Optimization to Remove Decode